Split responsibilities of User in to smaller classes
The User class had many responsibilities in the initial version. This
class now just depends on the Score and Progress classes to do the work.
Top level User specs could be deleted now because the functionality is
tested at lower levels. Also, found a bug in logic to accumulate the
progress points once the highest level is reached.